Extract and fix query tokens test
