+++ src/lib.rs
@@ -33,6 +33,7 @@ mod unittest_query_minters;
 mod unittest_query_all_tokens;
 mod unittest_query_nft_dossier;
 mod unittest_query_num_tokens;
+mod unittest_query_tokens;
 mod unittest_register_receive_nft;
 mod unittest_send;
 mod unittest_set_global_approval;
